[IMP] product_multi_barcode: split py files
