+++ product_multi_barcode/models/__init__.py
@@ -1,3 +1,5 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
+from . import product_barcode
+from . import product_template
 from . import product_product
